pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stm-gnn"
version = "1.0.0"
description = "STM-Graph: A Python Framework for Spatio-Temporal Mapping and Graph Neural Network Predictions"
readme = "README.md"
requires-python = ">=3.8,<3.9"
license = {text = "MIT"}
authors = [
    {name = "Amirhossein Ghaffari", email = "amirhossein.ghaffari@oulu.fi"}
]
keywords = ["spatial-temporal", "graph neural networks", "mapping", "torch geometric temporal", "framework"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "contextily>=1.5.2",
    "geopandas>=0.13.2",
    "matplotlib>=3.7.5",
    "networkx>=3.1",
    "numpy>=1.24.4",
    "osmnx==1.9.4",
    "pandas>=1.3.5",
    "pyproj>=3.5.0",
    "scikit_learn>=1.3.2",
    "scipy>=1.10.1",
    "Shapely>=2.0.6",
    "tqdm>=4.66.5",
    "wandb==0.19.1",
]

[project.optional-dependencies]
torch-extensions = [
    "torch_scatter>=2.1.2",
    "torch_sparse>=0.6.18",
    "torch_cluster>=1.6.3",
    "torch_spline_conv>=1.2.2",
    "torch_geometric==2.4.0",
    "torch_geometric_temporal==0.54.0",
]

[project.urls]
"Repository" = "https://github.com/Ahghaffari/stm_graph"
"Bug Tracker" = "https://github.com/Ahghaffari/stm_graph/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}

[tool.pip]
extra-index-url = ["https://download.pytorch.org/whl/cu118"]

